+++ src/App.tsx
@@ -64,10 +64,10 @@ const theme: ThemeProviderProps["theme"] = {
       light: "#f3e8ff",
     },
   },
-  iframe: {
-    fontFamily: "Avenir",
-    headerSize: "12px",
-  },
+  // iframe: {
+  //   fontFamily: "Avenir",
+  //   headerSize: "12px",
+  // },
 };
 
 // Glossary for translation
@@ -124,13 +124,7 @@ const components: DemoComponent[] = [
   {
     name: "zap",
     render: () => {
-      return (
-        <ZusAggregatedProfile
-          title="ZAP"
-          includePatientRecordSearch
-          includePatientDemographicsForm={false}
-        />
-      );
+      return <ZusAggregatedProfile title="ZAP" includePatientDemographicsForm={false} />;
     },
   },
   {
@@ -170,18 +164,14 @@ const DemoApp = ({ accessToken = "" }) => (
     }}
   >
     <PatientProvider patientID={VITE_PATIENT_ID} systemURL={VITE_SYSTEM_URL}>
-      <div className="App">
-        <h1>CTW Component Library</h1>
-
-        {demoComponents.map((demo, index) => (
-          <div className="ctw-space-y-5 ctw-bg-white ctw-p-1" key={index}>
-            <h3>
-              {demo.title} <small>{demo.note}</small>
-            </h3>
-            <ErrorBoundary>{demo.render()}</ErrorBoundary>
-          </div>
-        ))}
-      </div>
+      {demoComponents.map((demo, index) => (
+        <div className="ctw-space-y-5 ctw-bg-white ctw-p-1" key={index}>
+          <h3>
+            {demo.title} <small>{demo.note}</small>
+          </h3>
+          <ErrorBoundary>{demo.render()}</ErrorBoundary>
+        </div>
+      ))}
     </PatientProvider>
   </CTWProvider>
 );