+++ src/components/content/medications/helpers/details.tsx
@@ -1,21 +1,26 @@
 import { useMedicationHistoryEntries } from "./history";
+import { History } from "../../resource/helpers/history";
 import { useResourceDetailsDrawer } from "../../resource/resource-details-drawer";
 import { entryFromArray } from "@/components/core/data-list";
 import { Loading } from "@/components/core/loading";
-import { RowActionsProp } from "@/components/core/table/table-rows";
+import { RowActionsConfigProp } from "@/components/core/table/table-rows";
 import { useLastPrescriber } from "@/fhir/medications";
 import { MedicationStatementModel } from "@/fhir/models";
 
+const medicationHistory = (m: MedicationStatementModel) => (
+  <History getHistory={useMedicationHistoryEntries} model={m} />
+);
+
 export const useMedicationDetailsDrawer = ({
-  RowActions,
+  rowActions,
   enableDismissAndReadActions,
 }: {
-  RowActions?: RowActionsProp<MedicationStatementModel>;
+  rowActions?: (m: MedicationStatementModel) => RowActionsConfigProp<MedicationStatementModel>;
   enableDismissAndReadActions?: boolean;
 }) =>
   useResourceDetailsDrawer({
     header: (medication: MedicationStatementModel) => medication.display,
-    getHistory: useMedicationHistoryEntries,
+    renderChild: medicationHistory,
     details: (medication: MedicationStatementModel) => [
       { label: "Status", value: medication.displayStatus },
       { label: "Last Fill Date", value: medication.lastFillDate },
@@ -30,7 +35,7 @@ export const useMedicationDetailsDrawer = ({
       { label: "Last Prescribed Date", value: medication.lastPrescribedDate },
       ...entryFromArray("Note", medication.notesDisplay),
     ],
-    RowActions,
+    rowActions,
     enableDismissAndReadActions,
   });