+++ src/components/content/resource/resource-details-drawer.tsx
@@ -1,6 +1,5 @@
+import cx from "classnames";
 import { ReactNode, useEffect, useState } from "react";
-import { History, HistoryEntries } from "./helpers/history";
-import { Notes } from "./helpers/notes";
 import { useAdditionalResourceActions } from "./use-additional-resource-actions";
 import { DocumentButton } from "../CCDA/document-button";
 import { useCCDAModal } from "../CCDA/modal-ccda";
@@ -9,18 +8,15 @@ import { Drawer } from "@/components/core/drawer";
 import { Loading } from "@/components/core/loading";
 import { useAnalytics } from "@/components/core/providers/analytics/use-analytics";
 import { useDrawer } from "@/components/core/providers/drawer-provider";
+import { useCtwThemeRef } from "@/components/core/providers/theme/use-ctw-theme-ref";
 import { useCTW } from "@/components/core/providers/use-ctw";
-import { RowActionsProp } from "@/components/core/table/table-rows";
+import { RowActionsConfigProp } from "@/components/core/table/table-rows";
 import { getBinaryId } from "@/fhir/binaries";
 import { DocumentModel } from "@/fhir/models/document";
-import { EncounterModel } from "@/fhir/models/encounter";
 import { FHIRModel } from "@/fhir/models/fhir-model";
 import { searchProvenances } from "@/fhir/provenance";
+import { useBreakpoints } from "@/hooks/use-breakpoints";
 import { useFQSFeatureToggle } from "@/hooks/use-feature-toggle";
-import { sortBy } from "@/utils/nodash";
-import { UseQueryResultBasic } from "@/utils/request";
-
-const HISTORY_PAGE_LIMIT = 20;
 
 export type UseResourceDetailsDrawerProps<T extends fhir4.Resource, M extends FHIRModel<T>> = Pick<
   ResourceDetailsDrawerProps<T, M>,
@@ -28,9 +24,9 @@ export type UseResourceDetailsDrawerProps<T extends fhir4.Resource, M extends FH
   | "subHeader"
   | "getSourceDocument"
   | "details"
-  | "getHistory"
-  | "RowActions"
+  | "rowActions"
   | "enableDismissAndReadActions"
+  | "renderChild"
 >;
 
 export function useResourceDetailsDrawer<T extends fhir4.Resource, M extends FHIRModel<T>>(
@@ -51,36 +47,37 @@ export function useResourceDetailsDrawer<T extends fhir4.Resource, M extends FHI
 type ResourceDetailsDrawerProps<T extends fhir4.Resource, M extends FHIRModel<T>> = {
   className?: string;
   details: (model: M) => DetailsProps["details"];
-  getHistory?: (model: M) => UseQueryResultBasic<HistoryEntries | undefined>;
   getSourceDocument?: boolean;
   header: (model: M) => ReactNode;
   isOpen: boolean;
   model: M;
   onClose: () => void;
-  RowActions?: RowActionsProp<M>;
+  rowActions?: (model: M) => RowActionsConfigProp<M>;
   enableDismissAndReadActions?: boolean;
   subHeader?: (model: M) => ReactNode;
+  renderChild?: (model: M) => ReactNode;
 };
 
 function ResourceDetailsDrawer<T extends fhir4.Resource, M extends FHIRModel<T>>({
   className,
   details,
-  getHistory,
   getSourceDocument,
   header,
   isOpen,
   model,
   onClose,
-  RowActions,
+  rowActions,
   enableDismissAndReadActions,
   subHeader,
+  renderChild,
 }: ResourceDetailsDrawerProps<T, M>) {
   const openCCDAModal = useCCDAModal();
+  const ctwThemeRef = useCtwThemeRef();
   const [isLoading, setIsLoading] = useState(false);
   const [binaryId, setBinaryId] = useState<string>();
   const { getRequestContext } = useCTW();
   const fqsProvenances = useFQSFeatureToggle("provenances");
-  const history = getHistory && getHistory(model);
+  const breakpoints = useBreakpoints(ctwThemeRef);
 
   // We optionally look for any associated binary CCDAs
   // if getSourceDocument is true.
@@ -102,22 +99,28 @@ function ResourceDetailsDrawer<T extends fhir4.Resource, M extends FHIRModel<T>>
   }, [getSourceDocument, model, getRequestContext, fqsProvenances.enabled, fqsProvenances.ready]);
 
   const rowActionsWithAdditions = useAdditionalResourceActions({
-    RowActions,
+    rowActions,
     enableDismissAndReadActions,
+    isInFooter: true,
   });
 
   // We call rowActions right away so we'll know if it returns null and thus we should
   // hide our footer.
-  const actions =
-    rowActionsWithAdditions && rowActionsWithAdditions({ record: model, onSuccess: onClose });
+  const actions = rowActionsWithAdditions({
+    record: model,
+    onSuccess: onClose,
+    stacked: breakpoints.xs,
+  });
   const { trackInteraction } = useAnalytics();
 
   return (
     <Drawer className={className} title={model.resourceTypeTitle} isOpen={isOpen} onClose={onClose}>
       <Drawer.Body>
         <div className="ctw-space-y-4">
-          <div className="ctw-space-y-2">
-            <div className="ctw-text-3xl">{header(model)}</div>
+          <div className="ctw-space-y-2 ctw-px-2">
+            <div className={cx(breakpoints.xs ? "ctw-text-2xl" : "ctw-text-3xl")}>
+              {header(model)}
+            </div>
             {subHeader && <div>{subHeader(model)}</div>}
           </div>
 
@@ -141,21 +144,7 @@ function ResourceDetailsDrawer<T extends fhir4.Resource, M extends FHIRModel<T>>
               }
             />
           )}
-
-          {history &&
-            (history.isLoading ? (
-              <Loading message="Loading history..." />
-            ) : (
-              <History
-                entries={history.data ?? []}
-                limit={HISTORY_PAGE_LIMIT}
-                resourceTypeTitle={model.resourceTypeTitle}
-              />
-            ))}
-
-          {model instanceof EncounterModel && model.clinicalNotes.length > 0 && (
-            <Notes entries={sortBy(model.clinicalNotes, "title")} />
-          )}
+          {renderChild && renderChild(model)}
         </div>
       </Drawer.Body>
       {actions && (